Fix a harmless compiler warning in the sessions test harness.
FossilOrigin-Name: d389e20ab07a76d0c4be3801265977fa977a4394

+++ ext/session/test_session.c
@@ -253,7 +253,7 @@ static int test_obj_eq_string(Tcl_Obj *p, const char *z){
   int nObj;
   char *zObj;
 
-  n = strlen(z);
+  n = (int)strlen(z);
   zObj = Tcl_GetStringFromObj(p, &nObj);
 
   return (nObj==n && (n==0 || 0==memcmp(zObj, z, n)));
